mem0fun/matrixmem/requirements.txt
matrix-nio==0.25.2
mem0ai==1.0.11
ollama==0.6.1
qdrant-client==1.17.1
kuzu==0.11.3
rank-bm25==0.2.2
httpx==0.28.1
python-dotenv==1.2.2
pytest==9.0.2
